added Diamond Square Algorithm to calculate elevations displays Elevations map size has to be an  odd number e.g. (33 x 33, 257, 257)

+++ Biome.hpp
@@ -8,7 +8,7 @@ public:
     Biome(const Biome& orig);
     virtual ~Biome();
     int elevation;
-    void setBiomeAttributes(int biomeType);
+    void setBiomeAttributes(float biomeType);
     Tile getTile();
     
 private:
